clearing the cache
